fix. PostPage 컴포넌트 단위로 분리
