+++ src/components/PostContent.css
@@ -1,0 +1,100 @@
+/* PostContent.css */
+
+/* General Section Styling */
+.post_section {
+    margin-bottom: 30px; /* Space between sections */
+}
+
+.section_paragraph {
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 24px;
+    color: #333333;
+    margin-bottom: 16px; /* Space between paragraphs */
+}
+
+.section_paragraph .highlight {
+    font-weight: 600; /* Bold for highlighted text */
+}
+
+.section_heading {
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 600;
+    font-size: 20px;
+    line-height: 28px;
+    color: #000000;
+    margin-bottom: 20px;
+}
+
+ul {
+    list-style: disc; /* Default disc for unordered lists */
+    padding-left: 20px; /* Indent list items */
+    margin-bottom: 16px;
+}
+
+li {
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 24px;
+    color: #333333;
+    margin-bottom: 8px; /* Space between list items */
+}
+
+/* Code Block */
+.code_block {
+    background: #F8F8F8; /* Light grey background for code */
+    border-radius: 6px;
+    padding: 16px;
+    overflow-x: auto; /* Enable horizontal scrolling for long lines */
+    margin-bottom: 20px;
+}
+
+.code_block pre {
+    margin: 0; /* Remove default pre margin */
+}
+
+.code_block code {
+    font-family: 'D2 coding', 'Fira Code', monospace; /* Monospace font for code */
+    font-size: 14px;
+    line-height: 20px;
+    color: #000000;
+}
+
+/* Tip Section */
+.tip_section {
+    background: #F8F8F8; /* Light grey background */
+    border-left: 4px solid #626FE4; /* Blue left border */
+    border-radius: 6px;
+    padding: 20px;
+    margin-top: 30px;
+    margin-bottom: 30px;
+}
+
+.tip_title {
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 600;
+    font-size: 16px;
+    line-height: 24px;
+    color: #000000;
+    margin-bottom: 10px;
+}
+
+.tip_label {
+    color: #626FE4; /* Blue color for "Tip:" label */
+}
+
+.tip_content {
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 24px;
+    color: #333333;
+    margin-bottom: 16px;
+}
