+++ src/components/commentList.css
@@ -1,0 +1,7 @@
+/* CommentList.css */
+
+.comment_list {
+    /* Styles for the container of all comments */
+    margin-top: 20px; /* Space above the comment list */
+    /* Add any other specific styles for the list container */
+}
