+++ src/components/postContent.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+import "./PostContent.css";
+
+export default function PostContent({post}) {
+    return (
+        <>
+            <div className="post_section">
+              {post.content}
+            </div>
+        </>
+    )
+}
