+++ src/components/postFooter.css
@@ -1,0 +1,7 @@
+/* PostFooter.css */
+
+.post_footer {
+    margin-top: 30px;
+    padding-top: 20px;
+    border-top: 1px solid #EBEBEA; /* Separator line */
+}
