+++ src/components/postHeader.css
@@ -1,0 +1,141 @@
+/* PostHeader.css */
+
+.post_header {
+    margin-bottom: 30px;
+}
+
+.post_date {
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 14px;
+    line-height: 20px;
+    color: #8C8C8C; /* Lighter color for date */
+    margin-bottom: 8px;
+    display: block; /* Make it a block to take full width */
+}
+
+.post_title {
+    font-family: 'Pretendard';
+    font-style: normal;
+    font-weight: 700;
+    font-size: 28px;
+    line-height: 36px;
+    color: #000000;
+    margin: 0; /* Reset default heading margin */
+}
+
+/* Container for the entire bar */
+.profile_info_bar {
+    display: flex;
+    justify-content: space-between; /* Pushes content to the ends */
+    align-items: center; /* Vertically centers items */
+    padding: 20px 0; /* Add some vertical padding */
+    margin-bottom: 20px; /* Space below this bar */
+    border-bottom: 1px solid #EBEBEA; /* A subtle separator line */
+}
+
+/* Left side: Profile image and text */
+.profile_details {
+    display: flex;
+    align-items: center;
+    gap: 10px; /* Space between avatar and text */
+}
+
+.profile_avatar {
+    width: 40px; /* Size of the profile image */
+    height: 40px; /* Height of the profile image */
+    border-radius: 50%; /* Makes the image circular */
+    border: 1px solid #EBEBEA; /* Optional: subtle border around avatar */
+    object-fit: cover; /* Ensures image covers the area without distortion */
+}
+
+.posted_by_text {
+    font-family: 'Pretendard', sans-serif;
+    font-size: 14px;
+    line-height: 20px;
+    color: #8C8C8C; /* Grey color for general text */
+}
+
+.author_name {
+    font-weight: 600; /* Bold the author's name */
+    color: #333333; /* Darker color for the author's name */
+}
+
+/* Right side: Bell icon and toggle switch */
+.actions_right {
+    display: flex;
+    align-items: center;
+    gap: 15px; /* Space between bell icon and toggle */
+}
+
+.notification_icon {
+    width: 24px; /* Size of the bell icon */
+    height: 24px; /* Height of the bell icon */
+    cursor: pointer; /* Indicates it's clickable */
+}
+
+/* Toggle Switch Styling */
+.toggle_switch {
+    position: relative;
+    display: inline-block;
+    width: 40px; /* Width of the switch container */
+    height: 24px; /* Height of the switch container */
+    cursor: pointer;
+}
+
+/* Hide default HTML checkbox */
+.toggle_switch input {
+    opacity: 0;
+    width: 0;
+    height: 0;
+}
+
+/* The post_slider (the track) */
+.post_slider {
+    position: absolute;
+    cursor: pointer;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: #ccc; /* Default background color (off state) */
+    -webkit-transition: .4s;
+    transition: .4s;
+}
+
+/* The round slider handle */
+.post_slider:before {
+    position: absolute;
+    content: "";
+    height: 16px; /* Height of the handle */
+    width: 16px; /* Width of the handle */
+    left: 4px; /* Initial position from left */
+    bottom: 4px; /* Initial position from bottom */
+    background-color: white; /* Color of the handle */
+    -webkit-transition: .4s;
+    transition: .4s;
+}
+
+input:checked+.post_slider {
+    background-color: #626FE4; /* Background color when checked (on state) */
+}
+
+input:focus+.post_slider {
+    box-shadow: 0 0 1px #626FE4; /* Optional: focus effect */
+}
+
+input:checked+.post_slider:before {
+    -webkit-transform: translateX(16px); /* Move handle to the right */
+    -ms-transform: translateX(16px);
+    transform: translateX(16px);
+}
+
+/* Rounded post_sliders */
+.post_slider.round {
+    border-radius: 24px; /* Make the track rounded */
+}
+
+.post_slider.round:before {
+    border-radius: 50%; /* Make the handle circular */
+}
